% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_argos.R
\name{read_diag}
\alias{read_diag}
\title{Read Argos DIAG streams}
\usage{
read_diag(x)
}
\arguments{
\item{x}{DIAG connect (or character vector of files)}
}
\value{
dplry tbl_df
}
\description{
Read data from Argos PRV.
}
\details{
DIAG contains two sets of alternate coordinates,
}

