% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_argos.R
\name{read_argos}
\alias{read_argos}
\title{Read Argos PRV records.}
\usage{
read_argos(x, mungefix = TRUE, dtFormat = "\%Y-\%m-\%d \%H:\%M:\%S",
  tz = "UTC", duplicateTime_eps = 0.01,
  p4 = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0",
  verbose = FALSE)
}
\arguments{
\item{x}{Connection to PRV records}

\item{mungefix}{logical, should all track-validity errors be fixed.}

\item{dtFormat}{}

\item{tz}{}

\item{duplicateTime_eps}{}

\item{p4}{}

\item{verbose}{}
}
\value{
A dplyr tbl_df
}
\description{
Read data from Argos PRV.
}
\details{
See \code{\link{track_validity}} for details on the tests and fixes applied to records.
}

